LayoutTests/platform/chromium/storage/sql-error-codes-expected.txt

CONSOLE MESSAGE: line 37: Uncaught Exception thrown in transaction callback.
CONSOLE MESSAGE: line 69: Uncaught Cannot call toString() on this object.
This test tests the error codes reported in exceptional situations.
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code SYNTAX_ERR
PASS: expected and got error code UNKNOWN_ERR
PASS: expected and got error code QUOTA_ERR
PASS: expected and got error code VERSION_ERR
